More eslint fixes
